Fix key_down C# file order and Python color naming
