+++ public/usage-examples/input/key_down-1-example-oop.cs
@@ -1,55 +1,63 @@
-using SplashKitSDK;
-using static SplashKitSDK.SplashKit;
+ using SplashKitSDK;
 
-OpenWindow("Live Key Press Display", 800, 400);
-
-while (!QuitRequested())
+namespace KeyDownExample
 {
-    ProcessEvents();
-
-    bool spacePressed = KeyDown(KeyCode.SpaceKey);
-    bool leftPressed = KeyDown(KeyCode.LeftKey);
-    bool rightPressed = KeyDown(KeyCode.RightKey);
-    bool upPressed = KeyDown(KeyCode.UpKey);
-    bool downPressed = KeyDown(KeyCode.DownKey);
-
-    ClearScreen(ColorWhite());
-
-    DrawText("Hold the keys to see their current state.", ColorBlack(), 20, 20);
-
-    DrawKeyState("Space", spacePressed, 20, 80);
-    DrawKeyState("Left", leftPressed, 20, 130);
-    DrawKeyState("Right", rightPressed, 20, 180);
-    DrawKeyState("Up", upPressed, 20, 230);
-    DrawKeyState("Down", downPressed, 20, 280);
-
-    RefreshScreen(60);
-}
-
-CloseAllWindows();
-
-void DrawKeyState(string keyName, bool isPressed, double x, double y)
-{
-    Color circleColor;
-
-    if (isPressed)
+    public static class Program
     {
-        circleColor = ColorGreen();
-    }
-    else
-    {
-        circleColor = ColorGray();
-    }
+        public static void Main()
+        {
+            Window window = new Window("Live Key Press Display", 800, 400);
 
-    FillCircle(circleColor, x + 15, y + 15, 15);
-    DrawCircle(ColorBlack(), x + 15, y + 15, 15);
+            while (!SplashKit.QuitRequested())
+            {
+                SplashKit.ProcessEvents();
 
-    if (isPressed)
-    {
-        DrawText(keyName + ": Pressed", ColorBlack(), x + 40, y);
-    }
-    else
-    {
-        DrawText(keyName + ": Released", ColorBlack(), x + 40, y);
+                bool spacePressed = SplashKit.KeyDown(KeyCode.SpaceKey);
+                bool leftPressed = SplashKit.KeyDown(KeyCode.LeftKey);
+                bool rightPressed = SplashKit.KeyDown(KeyCode.RightKey);
+                bool upPressed = SplashKit.KeyDown(KeyCode.UpKey);
+                bool downPressed = SplashKit.KeyDown(KeyCode.DownKey);
+
+                window.Clear(Color.White);
+
+                window.DrawText("Hold the keys to see their current state.", Color.Black, 20, 20);
+
+                DrawKeyState(window, "Space", spacePressed, 20, 80);
+                DrawKeyState(window, "Left", leftPressed, 20, 130);
+                DrawKeyState(window, "Right", rightPressed, 20, 180);
+                DrawKeyState(window, "Up", upPressed, 20, 230);
+                DrawKeyState(window, "Down", downPressed, 20, 280);
+
+                window.Refresh(60);
+            }
+
+            SplashKit.CloseAllWindows();
+        }
+
+        public static void DrawKeyState(Window window, string keyName, bool isPressed, double x, double y)
+        {
+            Color circleColor;
+
+            if (isPressed)
+            {
+                circleColor = Color.Green;
+            }
+            else
+            {
+                circleColor = Color.Gray;
+            }
+
+            window.FillCircle(circleColor, x + 15, y + 15, 15);
+            window.DrawCircle(Color.Black, x + 15, y + 15, 15);
+
+            if (isPressed)
+            {
+                window.DrawText(keyName + ": Pressed", Color.Black, x + 40, y);
+            }
+            else
+            {
+                window.DrawText(keyName + ": Released", Color.Black, x + 40, y);
+            }
+        }
     }
 }
