+++ public/usage-examples/input/key_down-1-example.py
@@ -3,17 +3,17 @@ from splashkit import *
 
 def draw_key_state(key_name, is_pressed, x, y):
     if is_pressed:
-        circle_color = Color_green()
+        circle_color = color_green()
     else:
-        circle_color = Color_gray()
+        circle_color = color_gray()
 
     fill_circle(circle_color, x + 15, y + 15, 15)
-    draw_circle(Color_black(), x + 15, y + 15, 15)
+    draw_circle(color_black(), x + 15, y + 15, 15)
 
     if is_pressed:
-        draw_text(key_name + ": Pressed", Color_black(), x + 40, y)
+        draw_text(key_name + ": Pressed", color_black(), x + 40, y)
     else:
-        draw_text(key_name + ": Released", Color_black(), x + 40, y)
+        draw_text(key_name + ": Released", color_black(), x + 40, y)
 
 
 open_window("Live Key Press Display", 800, 400)
@@ -27,9 +27,9 @@ while not quit_requested():
     up_pressed = key_down(UP_KEY)
     down_pressed = key_down(DOWN_KEY)
 
-    clear_screen(Color_white())
+    clear_screen(color_white())
 
-    draw_text("Hold the keys to see their current state.", Color_black(), 20, 20)
+    draw_text("Hold the keys to see their current state.", color_black(), 20, 20)
 
     draw_key_state("Space", space_pressed, 20, 80)
     draw_key_state("Left", left_pressed, 20, 130)
